refactor(tests): split router tests into smaller chunks
